# Conflicts:
#	cl-robdd-scala/src/main/scala/genus/RandomType.scala
#	cl-robdd-scala/src/main/scala/genus/TypeSystemWithLBdd.scala
#	cl-robdd-scala/src/main/scala/genus/Types.scala
#	cl-robdd-scala/src/main/scala/rte/Combination.scala
#	cl-robdd-scala/src/main/scala/rte/Rte.scala
#	cl-robdd-scala/src/main/scala/xymbolyco/Thompson.scala
#	cl-robdd-scala/src/test/scala/xymbolyco/ThompsonTestSuite.scala

+++ cl-robdd-scala/src/main/scala/treereduce/RationalFoldTest.scala
@@ -33,7 +33,7 @@ object RationalFoldTest {
   def time1[R](block: => R):Double = {
     val t0 = System.nanoTime()
     block
-    System.nanoTime() - t0
+    (System.nanoTime() - t0).toDouble
   }
 
   def time[R](repetitions: Int, name: String, block: => R): Double = {
